# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "TodoBackEndAPI",
        "description": "TODO uygulaması CRUD işlerimleri örneği."
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all rows",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create row",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/EventItemModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get row by id",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update row by id",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "in": "path"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/EventItemModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete row by id",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            }
        },
        "/events/changeStatu/{id}": {
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Dynamic change statu by id",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "type": "integer",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "$ref": "#/definitions/EventResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "EventResponseModel": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/EventItemModel"
                    }
                }
            }
        },
        "EventItemModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "statu": {
                    "type": "integer"
                }
            }
        }
    }
}